Split AppDelegate notification handling
