+++ Timer/AppDelegate.swift
@@ -150,29 +150,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
     return true
   }
 
-  nonisolated func userNotificationCenter(
-    _: UNUserNotificationCenter,
-    willPresent _: UNNotification,
-    withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
-    completionHandler([.banner, .sound])
-  }
-
-  nonisolated func userNotificationCenter(
-    _: UNUserNotificationCenter,
-    didReceive response: UNNotificationResponse,
-    withCompletionHandler completionHandler: @escaping () -> Void) {
-    let actionIdentifier = response.actionIdentifier
-    let controllerIdentifier = response.notification.request.content.userInfo[
-      MVNotificationUserInfoKeys.controllerIdentifier
-    ] as? String
-
-    completionHandler()
-
-    DispatchQueue.main.async { [weak self] in
-      self?.handleNotificationAction(actionIdentifier, controllerIdentifier: controllerIdentifier)
-    }
-  }
-
   func addBadgeToDock(controller: MVTimerController) {
     if self.currentlyInDock != controller {
       self.removeBadgeFromDock()
@@ -211,29 +188,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
     self.staysOnTop = UserDefaults.standard.bool(forKey: MVUserDefaultsKeys.staysOnTop)
   }
 
-  private func handleNotificationAction(_ actionIdentifier: String, controllerIdentifier: String?) {
-    let controller = self.controller(matching: controllerIdentifier)
-
-    switch actionIdentifier {
-    case MVNotificationIdentifiers.restartTimerActionIdentifier:
-      controller.restartLastTimer()
-
-    case MVNotificationIdentifiers.addFiveMinutesActionIdentifier:
-      controller.addTime(seconds: CGFloat(5 * 60))
-
-    case MVNotificationIdentifiers.stopTimerActionIdentifier:
-      controller.stopTimer()
-
-    case UNNotificationDefaultActionIdentifier:
-      controller.window?.makeKeyAndOrderFront(nil)
-      NSApplication.shared.activate(ignoringOtherApps: true)
-
-    default:
-      break
-    }
-  }
-
-  private func controller(matching identifier: String?) -> MVTimerController {
+  func controller(matching identifier: String?) -> MVTimerController {
     if let identifier,
        let controller = self.controllers.first(where: { $0.identifier == identifier }) {
       return controller
@@ -248,32 +203,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
     self.controllers.append(controller)
     self.addBadgeToDock(controller: controller)
     return controller
-  }
-
-  private func registerNotificationCategories() {
-    let restartAction = UNNotificationAction(
-      identifier: MVNotificationIdentifiers.restartTimerActionIdentifier,
-      title: "Restart",
-      options: []
-    )
-    let addFiveMinutesAction = UNNotificationAction(
-      identifier: MVNotificationIdentifiers.addFiveMinutesActionIdentifier,
-      title: "+5 min",
-      options: []
-    )
-    let stopAction = UNNotificationAction(
-      identifier: MVNotificationIdentifiers.stopTimerActionIdentifier,
-      title: "Stop",
-      options: [.destructive]
-    )
-    let category = UNNotificationCategory(
-      identifier: MVNotificationIdentifiers.timerCompleteCategoryIdentifier,
-      actions: [restartAction, addFiveMinutesAction, stopAction],
-      intentIdentifiers: [],
-      options: []
-    )
-
-    UNUserNotificationCenter.current().setNotificationCategories([category])
   }
 
   private func observeNotifications() {
